flask_app/requirements.txt
aiohttp==3.8.6
aiohttp-cors==0.7.0
aiokafka==0.8.0
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
bidict==0.22.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorclass==2.2.2
colorlog==6.8.2
croniter==2.0.5
dnspython==2.3.0
eventlet==0.36.1
faust-streaming==0.10.13
Flask==2.2.5
Flask-SocketIO==5.3.6
flask-sqlalchemy==3.0.5
frozenlist==1.3.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib-metadata==6.7.0
intervaltree==3.1.0
itsdangerous==2.1.2
jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==2.1.5
mode==4.3.2
mode-streaming==0.3.5
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==1.4.6
opentracing==1.3.0
packaging==24.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-engineio==4.9.1
python-socketio==5.11.0
pytz==2024.1
simple-websocket==1.0.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
terminaltables==3.1.10
typing-extensions==4.7.1
venusian==3.0.0
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.9.4
zipp==3.15.0
